requirements.txt
torch
numpy
scipy
astropy
lightkurve
pymc
numpyro
arviz
matplotlib
pyyaml
hydra-core
tqdm
astroquery
requests
pandas
